Refactor routes related structure
