+++ frontend/svelte/src/lib/constants/routes.constants.ts
@@ -1,0 +1,9 @@
+export enum AppPath {
+  Authentication = "/",
+  Accounts = "/#/accounts",
+  Neurons = "/#/neurons",
+  Proposals = "/#/proposals",
+  Canisters = "/#/canisters",
+  Wallet = "/#/wallet",
+  ProposalDetails = "/#/proposal",
+}
