+++ frontend/svelte/src/routes/Accounts.svelte
@@ -9,8 +9,8 @@
   import { i18n } from "../lib/stores/i18n";
   import Toolbar from "../lib/components/ui/Toolbar.svelte";
   import Spinner from "../lib/components/ui/Spinner.svelte";
-  import { AppPath } from "./routes";
   import { routeStore } from "../lib/stores/route.store";
+  import { AppPath } from "../lib/constants/routes.constants";
 
   // TODO: To be removed once this page has been implemented
   onMount(() => {
